requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.1
attrs==23.2.0
awscli==1.29.54
backoff==2.2.1
blinker==1.8.1
boto3==1.28.23
botocore==1.31.54
cachetools==5.3.3
cattrs==23.2.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cmasher==1.8.0
colorama==0.4.4
coloredlogs==15.0.1
colorspacious==1.1.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.10.0
Cython==3.0.0
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
fastapi==0.110.3
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanfriendly==10.0
idna==2.10
imageio==2.34.1
inference==0.11.2
inference-sdk==0.9.23
inference-tools==0.13.0
inflect==7.2.1
iniconfig==2.0.0
intel-openmp==2021.4.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lapx==0.5.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mplsoccer==1.2.4
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.15.1
openai==1.30.5
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
packaging==24.0
pandas==2.2.2
pendulum==3.0.0
piexif==1.1.3
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
PuLP==2.8.0
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.3.2
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.0
pydot==2.0.0
Pygments==2.17.2
pyparsing==3.1.2
pyreadline3==3.4.1
pytest==8.2.1
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pywin32==306
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
requests-cache==1.2.0
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.28
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.6.2
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.1
six==1.16.0
skypilot==0.5.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
statsbombpy==1.13.0
streamlit==1.34.0
structlog==24.2.0
supervision==0.20.0
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.5.22
time-machine==2.14.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
typeguard==4.3.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.8
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.18
watchdog==4.0.1
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.9.4
zxing-cpp==2.2.0
